pyproject.toml
[tool.poetry]
name = "xeroapi"
version = "0.1.6"
description = ""
authors = ["tim-meiwald-bluefruit <tim.meiwald@bluefruit.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
email-validator = "^1.3.1"
httpx = "^0.23.3"


[tool.poetry.group.dev.dependencies]
fastapi = "^0.95.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.21.1"
email-validator = "^1.3.1"
httpx = "^0.23.3"
pytest = "^7.2.2"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
